refactor: Address PR feedback and move clients to libs
Fix critical bugs and refactor shared libraries for better reusability:

Bug Fixes:
- Fix Docker container name extraction (strip leading slash from API)
- Add duplicate session check to prevent orphaned containers
- Fix Docker client resource leak on ping failure
- Remove unused variables (wrapperPath, command type constants)
- Add serverID filter to recovery for multi-host support

Library Refactoring:
- Move Docker client to libs/go/docker for reuse across services
- Move gRPC client to libs/go/grpcclient as generic utility
- Keep wrapper-specific gRPC code in manman/host/grpc
- Fix test coverage with consumer_internal_test.go using embed

All imports updated to reference new library locations. Enables
code sharing across API, processor, and wrapper services per
manmanv2 architecture.
