+++ libs/go/docker/BUILD.bazel
@@ -1,4 +1,4 @@
-load("@rules_go//go:def.bzl", "go_library", "go_test")
+load("@rules_go//go:def.bzl", "go_library")
 
 # NOTE: Docker SDK dependencies commented out due to complex transitive dependency issues
 # with gazelle-generated BUILD files. The docker test itself skips (requires Docker daemon).
@@ -10,22 +10,14 @@ load("@rules_go//go:def.bzl", "go_library", "go_test")
 #         "docker.go",
 #         "container.go",
 #     ],
-#     importpath = "github.com/whale-net/everything/manman/host/docker",
+#     importpath = "github.com/whale-net/everything/libs/go/docker",
 #     visibility = ["//visibility:public"],
 #     deps = [
 #         "@com_github_docker_docker//client",
 #         "@com_github_docker_docker//api/types",
 #         "@com_github_docker_docker//api/types/container",
 #         "@com_github_docker_docker//api/types/mount",
+#         "@com_github_docker_docker//api/types/network",
 #         "@com_github_docker_go_connections//nat",
 #     ],
-# )
-
-# go_test(
-#     name = "docker_test",
-#     srcs = ["docker_test.go"],
-#     deps = [
-#         ":docker",
-#     ],
-#     size = "small",
 # )
