+++ libs/go/grpcclient/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "grpcclient",
+    srcs = ["client.go"],
+    importpath = "github.com/whale-net/everything/libs/go/grpcclient",
+    visibility = ["//visibility:public"],
+    deps = [
+        "@org_golang_google_grpc//:grpc",
+        "@org_golang_google_grpc//credentials/insecure",
+    ],
+)
