--- manman/host/docker/docker_test.go
@@ -1,27 +1,0 @@
-package docker_test
-
-import (
-	"context"
-	"testing"
-
-	"github.com/whale-net/everything/manman/host/docker"
-)
-
-// TestNewClient tests creating a new Docker client
-// This is a placeholder test - actual testing would require a Docker daemon
-func TestNewClient(t *testing.T) {
-	// Skip if Docker socket is not available
-	t.Skip("Requires Docker daemon")
-	
-	ctx := context.Background()
-	client, err := docker.NewClient("")
-	if err != nil {
-		t.Fatalf("Failed to create Docker client: %v", err)
-	}
-	defer client.Close()
-
-	// Test that we can ping Docker
-	// This would require access to the actual client
-	_ = ctx
-	_ = client
-}
