+++ manman/host/grpc/BUILD.bazel
@@ -1,25 +1,14 @@
-load("@rules_go//go:def.bzl", "go_library", "go_test")
+load("@rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "grpc",
     srcs = [
-        "client.go",
         "wrapper_control.go",
     ],
     importpath = "github.com/whale-net/everything/manman/host/grpc",
     visibility = ["//visibility:public"],
     deps = [
+        "//libs/go/grpcclient",
         "//manman/protos:manmanpb",
-        "@org_golang_google_grpc//:grpc",
-        "@org_golang_google_grpc//credentials/insecure",
     ],
-)
-
-go_test(
-    name = "grpc_test",
-    srcs = ["client_test.go"],
-    deps = [
-        ":grpc",
-    ],
-    size = "small",
 )
